requirements.txt

Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.1
pandas==1.0.1
scikit-image==0.16.2
tensorflow-gpu==2.1.0
